Initial upload for the hardware prototype
Nothing here is close to being finalized.  I have it running on a
breadboard with some successful output.  I still need to make the
schematic, order the pcbs and finalize the firmware.

I wanted to upload it now so I don't lose track even though I'm not
crazy about having this stuff public since it is all hacky at the
moment.